# apps/default/v1/workflow.openapi.yaml
openapi: 3.1.0
info:
  title: Trustage Workflow API
  description: Workflow definition management for the Trustage orchestrator.
  version: v1.0.0
x-service-namespace: service_trustage
x-service-permissions:
  - event_ingest
  - workflow_view
  - workflow_manage
  - instance_view
  - instance_retry
  - execution_view
  - execution_retry
  - execution_resume
  - signal_send
paths:
  /workflow.v1.WorkflowService/ActivateWorkflow:
    post:
      tags:
        - workflow.v1.WorkflowService
      summary: ActivateWorkflow
      operationId: workflow.v1.WorkflowService.ActivateWorkflow
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/workflow.v1.ActivateWorkflowRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/workflow.v1.ActivateWorkflowResponse'
      x-required-permissions:
        - workflow_manage
  /workflow.v1.WorkflowService/CreateWorkflow:
    post:
      tags:
        - workflow.v1.WorkflowService
      summary: CreateWorkflow
      operationId: workflow.v1.WorkflowService.CreateWorkflow
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/workflow.v1.CreateWorkflowRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/workflow.v1.CreateWorkflowResponse'
      x-required-permissions:
        - workflow_manage
  /workflow.v1.WorkflowService/GetWorkflow:
    post:
      tags:
        - workflow.v1.WorkflowService
      summary: GetWorkflow
      operationId: workflow.v1.WorkflowService.GetWorkflow
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/workflow.v1.GetWorkflowRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/workflow.v1.GetWorkflowResponse'
      x-required-permissions:
        - workflow_view
  /workflow.v1.WorkflowService/ListWorkflows:
    post:
      tags:
        - workflow.v1.WorkflowService
      summary: ListWorkflows
      operationId: workflow.v1.WorkflowService.ListWorkflows
      parameters:
        - name: Connect-Protocol-Version
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/connect-protocol-version'
        - name: Connect-Timeout-Ms
          in: header
          schema:
            $ref: '#/components/schemas/connect-timeout-header'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/workflow.v1.ListWorkflowsRequest'
        required: true
      responses:
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connect.error'
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/workflow.v1.ListWorkflowsResponse'
      x-required-permissions:
        - workflow_view
components:
  schemas:
    common.v1.PageCursor:
      type: object
      properties:
        limit:
          type: integer
          title: limit
          format: int32
          description: Number of items per cursor (limit)
        page:
          type: string
          title: page
          description: Page number/ cursor representing last id from result
      title: PageCursor
      additionalProperties: false
      description: |-
        PageCursor provides standard offset-based pagination/cursor stream parameters.
         Used for list operations that return large result sets.
    common.v1.SearchRequest:
      type: object
      properties:
        query:
          type: string
          title: query
          description: Full-text search query string
        idQuery:
          type: string
          title: id_query
          description: Specific ID or ID pattern to search for
        cursor:
          title: cursor
          description: Pagination parameters
          $ref: '#/components/schemas/common.v1.PageCursor'
        properties:
          type: array
          items:
            type: string
          title: properties
          description: Specific properties/fields to include in results
        extras:
          title: extras
          description: Service-specific additional search parameters
          $ref: '#/components/schemas/google.protobuf.Struct'
      title: SearchRequest
      additionalProperties: false
      description: |-
        SearchRequest provides a standard structure for search operations across services.
         Supports text search, ID-based queries, pagination, property filtering, and extensibility.
    connect-protocol-version:
      type: number
      title: Connect-Protocol-Version
      enum:
        - 1
      description: Define the version of the Connect protocol
      const: 1
    connect-timeout-header:
      type: number
      title: Connect-Timeout-Ms
      description: Define the timeout, in ms
    connect.error:
      type: object
      properties:
        code:
          type: string
          examples:
            - not_found
          enum:
            - canceled
            - unknown
            - invalid_argument
            - deadline_exceeded
            - not_found
            - already_exists
            - permission_denied
            - resource_exhausted
            - failed_precondition
            - aborted
            - out_of_range
            - unimplemented
            - internal
            - unavailable
            - data_loss
            - unauthenticated
          description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
        message:
          type: string
          description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
        details:
          type: array
          items:
            $ref: '#/components/schemas/connect.error_details.Any'
          description: A list of messages that carry the error details. There is no limit on the number of messages.
      title: Connect Error
      additionalProperties: true
      description: 'Error type returned by Connect: https://connectrpc.com/docs/go/errors/#http-representation'
    connect.error_details.Any:
      type: object
      properties:
        type:
          type: string
          description: 'A URL that acts as a globally unique identifier for the type of the serialized message. For example: `type.googleapis.com/google.rpc.ErrorInfo`. This is used to determine the schema of the data in the `value` field and is the discriminator for the `debug` field.'
        value:
          type: string
          format: binary
          description: The Protobuf message, serialized as bytes and base64-encoded. The specific message type is identified by the `type` field.
        debug:
          oneOf:
            - type: object
              title: Any
              additionalProperties: true
              description: Detailed error information.
          discriminator:
            propertyName: type
          title: Debug
          description: Deserialized error detail payload. The 'type' field indicates the schema. This field is for easier debugging and should not be relied upon for application logic.
      additionalProperties: true
      description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message, with an additional debug field for ConnectRPC error details.
    google.protobuf.ListValue:
      type: object
      properties:
        values:
          type: array
          items:
            $ref: '#/components/schemas/google.protobuf.Value'
          title: values
          description: Repeated field of dynamically typed values.
      title: ListValue
      additionalProperties: false
      description: |-
        `ListValue` is a wrapper around a repeated field of values.

         The JSON representation for `ListValue` is JSON array.
    google.protobuf.NullValue:
      type: string
      title: NullValue
      enum:
        - NULL_VALUE
      description: |-
        `NullValue` is a singleton enumeration to represent the null value for the
         `Value` type union.

         The JSON representation for `NullValue` is JSON `null`.
    google.protobuf.Struct:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/google.protobuf.Value'
      description: |-
        `Struct` represents a structured data value, consisting of fields
         which map to dynamically typed values. In some languages, `Struct`
         might be supported by a native representation. For example, in
         scripting languages like JS a struct is represented as an
         object. The details of that representation are described together
         with the proto support for the language.

         The JSON representation for `Struct` is JSON object.
    google.protobuf.Struct.FieldsEntry:
      type: object
      properties:
        key:
          type: string
          title: key
        value:
          title: value
          $ref: '#/components/schemas/google.protobuf.Value'
      title: FieldsEntry
      additionalProperties: false
    google.protobuf.Timestamp:
      type: string
      examples:
        - "2023-01-15T01:30:15.01Z"
        - "2024-12-25T12:00:00Z"
      format: date-time
      description: |-
        A Timestamp represents a point in time independent of any time zone or local
         calendar, encoded as a count of seconds and fractions of seconds at
         nanosecond resolution. The count is relative to an epoch at UTC midnight on
         January 1, 1970, in the proleptic Gregorian calendar which extends the
         Gregorian calendar backwards to year one.

         All minutes are 60 seconds long. Leap seconds are "smeared" so that no leap
         second table is needed for interpretation, using a [24-hour linear
         smear](https://developers.google.com/time/smear).

         The range is from 0001-01-01T00:00:00Z to 9999-12-31T23:59:59.999999999Z. By
         restricting to that range, we ensure that we can convert to and from [RFC
         3339](https://www.ietf.org/rfc/rfc3339.txt) date strings.

         # Examples

         Example 1: Compute Timestamp from POSIX `time()`.

             Timestamp timestamp;
             timestamp.set_seconds(time(NULL));
             timestamp.set_nanos(0);

         Example 2: Compute Timestamp from POSIX `gettimeofday()`.

             struct timeval tv;
             gettimeofday(&tv, NULL);

             Timestamp timestamp;
             timestamp.set_seconds(tv.tv_sec);
             timestamp.set_nanos(tv.tv_usec * 1000);

         Example 3: Compute Timestamp from Win32 `GetSystemTimeAsFileTime()`.

             FILETIME ft;
             GetSystemTimeAsFileTime(&ft);
             UINT64 ticks = (((UINT64)ft.dwHighDateTime) << 32) | ft.dwLowDateTime;

             // A Windows tick is 100 nanoseconds. Windows epoch 1601-01-01T00:00:00Z
             // is 11644473600 seconds before Unix epoch 1970-01-01T00:00:00Z.
             Timestamp timestamp;
             timestamp.set_seconds((INT64) ((ticks / 10000000) - 11644473600LL));
             timestamp.set_nanos((INT32) ((ticks % 10000000) * 100));

         Example 4: Compute Timestamp from Java `System.currentTimeMillis()`.

             long millis = System.currentTimeMillis();

             Timestamp timestamp = Timestamp.newBuilder().setSeconds(millis / 1000)
                 .setNanos((int) ((millis % 1000) * 1000000)).build();

         Example 5: Compute Timestamp from Java `Instant.now()`.

             Instant now = Instant.now();

             Timestamp timestamp =
                 Timestamp.newBuilder().setSeconds(now.getEpochSecond())
                     .setNanos(now.getNano()).build();

         Example 6: Compute Timestamp from current time in Python.

             timestamp = Timestamp()
             timestamp.GetCurrentTime()

         # JSON Mapping

         In JSON format, the Timestamp type is encoded as a string in the
         [RFC 3339](https://www.ietf.org/rfc/rfc3339.txt) format. That is, the
         format is "{year}-{month}-{day}T{hour}:{min}:{sec}[.{frac_sec}]Z"
         where {year} is always expressed using four digits while {month}, {day},
         {hour}, {min}, and {sec} are zero-padded to two digits each. The fractional
         seconds, which can go up to 9 digits (i.e. up to 1 nanosecond resolution),
         are optional. The "Z" suffix indicates the timezone ("UTC"); the timezone
         is required. A proto3 JSON serializer should always use UTC (as indicated by
         "Z") when printing the Timestamp type and a proto3 JSON parser should be
         able to accept both UTC and other timezones (as indicated by an offset).

         For example, "2017-01-15T01:30:15.01Z" encodes 15.01 seconds past
         01:30 UTC on January 15, 2017.

         In JavaScript, one can convert a Date object to this format using the
         standard
         [toISOString()](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/toISOString)
         method. In Python, a standard `datetime.datetime` object can be converted
         to this format using
         [`strftime`](https://docs.python.org/2/library/time.html#time.strftime) with
         the time format spec '%Y-%m-%dT%H:%M:%S.%fZ'. Likewise, in Java, one can use
         the Joda Time's [`ISODateTimeFormat.dateTime()`](
         http://joda-time.sourceforge.net/apidocs/org/joda/time/format/ISODateTimeFormat.html#dateTime()
         ) to obtain a formatter capable of generating timestamps in this format.
    google.protobuf.Value:
      oneOf:
        - type: "null"
        - type: number
        - type: string
        - type: boolean
        - type: array
        - type: object
          additionalProperties: true
      description: |-
        `Value` represents a dynamically typed value which can be either
         null, a number, a string, a boolean, a recursive struct value, or a
         list of values. A producer of value is expected to set one of these
         variants. Absence of any variant indicates an error.

         The JSON representation for `Value` is JSON value.
    workflow.v1.ActivateWorkflowRequest:
      type: object
      properties:
        id:
          type: string
          title: id
      title: ActivateWorkflowRequest
      additionalProperties: false
    workflow.v1.ActivateWorkflowResponse:
      type: object
      properties:
        workflow:
          title: workflow
          $ref: '#/components/schemas/workflow.v1.WorkflowDefinition'
      title: ActivateWorkflowResponse
      additionalProperties: false
    workflow.v1.CreateWorkflowRequest:
      type: object
      properties:
        dsl:
          title: dsl
          $ref: '#/components/schemas/google.protobuf.Struct'
      title: CreateWorkflowRequest
      additionalProperties: false
    workflow.v1.CreateWorkflowResponse:
      type: object
      properties:
        workflow:
          title: workflow
          $ref: '#/components/schemas/workflow.v1.WorkflowDefinition'
      title: CreateWorkflowResponse
      additionalProperties: false
    workflow.v1.GetWorkflowRequest:
      type: object
      properties:
        id:
          type: string
          title: id
      title: GetWorkflowRequest
      additionalProperties: false
    workflow.v1.GetWorkflowResponse:
      type: object
      properties:
        workflow:
          title: workflow
          $ref: '#/components/schemas/workflow.v1.WorkflowDefinition'
        schedules:
          type: array
          items:
            $ref: '#/components/schemas/workflow.v1.ScheduleDefinition'
          title: schedules
      title: GetWorkflowResponse
      additionalProperties: false
    workflow.v1.ListWorkflowsRequest:
      type: object
      properties:
        name:
          type: string
          title: name
        status:
          title: status
          $ref: '#/components/schemas/workflow.v1.WorkflowStatus'
        search:
          title: search
          $ref: '#/components/schemas/common.v1.SearchRequest'
      title: ListWorkflowsRequest
      additionalProperties: false
    workflow.v1.ListWorkflowsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/workflow.v1.WorkflowDefinition'
          title: items
        nextCursor:
          title: next_cursor
          $ref: '#/components/schemas/common.v1.PageCursor'
      title: ListWorkflowsResponse
      additionalProperties: false
    workflow.v1.ScheduleDefinition:
      type: object
      properties:
        id:
          type: string
          title: id
        name:
          type: string
          title: name
        cronExpr:
          type: string
          title: cron_expr
        workflowName:
          type: string
          title: workflow_name
        workflowVersion:
          type: integer
          title: workflow_version
          format: int32
        active:
          type: boolean
          title: active
        nextFireAt:
          title: next_fire_at
          $ref: '#/components/schemas/google.protobuf.Timestamp'
        lastFiredAt:
          title: last_fired_at
          $ref: '#/components/schemas/google.protobuf.Timestamp'
        jitterSeconds:
          type: integer
          title: jitter_seconds
          format: int32
        createdAt:
          title: created_at
          $ref: '#/components/schemas/google.protobuf.Timestamp'
        updatedAt:
          title: updated_at
          $ref: '#/components/schemas/google.protobuf.Timestamp'
      title: ScheduleDefinition
      additionalProperties: false
    workflow.v1.WorkflowDefinition:
      type: object
      properties:
        id:
          type: string
          title: id
        name:
          type: string
          title: name
        version:
          type: integer
          title: version
          format: int32
        status:
          title: status
          $ref: '#/components/schemas/workflow.v1.WorkflowStatus'
        dsl:
          title: dsl
          $ref: '#/components/schemas/google.protobuf.Struct'
        inputSchemaHash:
          type: string
          title: input_schema_hash
        timeoutSeconds:
          type:
            - integer
            - string
          title: timeout_seconds
          format: int64
        createdAt:
          title: created_at
          $ref: '#/components/schemas/google.protobuf.Timestamp'
        updatedAt:
          title: updated_at
          $ref: '#/components/schemas/google.protobuf.Timestamp'
      title: WorkflowDefinition
      additionalProperties: false
    workflow.v1.WorkflowStatus:
      type: string
      title: WorkflowStatus
      enum:
        - WORKFLOW_STATUS_UNSPECIFIED
        - WORKFLOW_STATUS_DRAFT
        - WORKFLOW_STATUS_ACTIVE
        - WORKFLOW_STATUS_ARCHIVED
security: []
tags:
  - name: workflow.v1.WorkflowService